+++ CHANGELOG.md
@@ -107,6 +107,9 @@
   that contain one of the patterns.  
   [kasrababaei](https://github.com/kasrababaei)
 
+* Make `empty_count` auto-correctable.  
+  [KS1019](https://github.com/KS1019/)
+
 #### Bug Fixes
 
 * Silence `discarded_notification_center_observer` rule in closures. Furthermore,